safari-600.7-branch/LayoutTests/inspector-obsolete/styles/inject-stylesheet-expected.txt

CONSOLE MESSAGE: line 9: iframe loaded
Tests that injected user stylesheets are reflected in the Styles pane.


Main frame style:
[expanded] 
-webkit-background-clip: border-box;
-webkit-background-origin: padding-box;
-webkit-background-size: auto;
-webkit-border-image: url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAQAAAC1+jfqAAAAiElEQ…Ms+LS30CAhBN5nNxeT5hbJ1zwmji2k+aF6NENIPf/hs54f0sZFUVAMigAAAABJRU5ErkJggg==) 100% / 1 / 0px stretch;
    #main - url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAQAAAC1+jfqAAAAiElEQVR42r2RsQrDMAxEBRdl8SDcX8lQPGg1GBI6lvz/h7QyRRXV0qUULwfvwZ1tenw5PxToRPWMC52eA9+WDnlh3HFQ/xBQl86NFYJqeGflkiogrOvVlIFhqURFVho3x1moGAa3deMs+LS30CAhBN5nNxeT5hbJ1zwmji2k+aF6NENIPf/hs54f0sZFUVAMigAAAABJRU5ErkJggg==) user stylesheet
background-attachment: scroll;
    #main - initial inject-stylesheet.html:5
background-clip: border-box;
    #main - initial inject-stylesheet.html:5
background-color: rgb(0, 0, 255);
    #main - blue inject-stylesheet.html:5
background-image: none;
    #main - initial inject-stylesheet.html:5
background-origin: padding-box;
    #main - initial inject-stylesheet.html:5
background-size: auto;
    #main - initial inject-stylesheet.html:5
color: rgb(255, 0, 0);
    #main - red user stylesheet
display: block;
    div - block user agent stylesheet

[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
media="screen" (inject-stylesheet.html)
#main  { (inject-stylesheet.html:5)
background: blue;
    background-image: initial;
    background-position-x: initial;
    background-position-y: initial;
    background-size: initial;
    background-repeat-x: initial;
    background-repeat-y: initial;
    background-attachment: initial;
    background-origin: initial;
    background-clip: initial;
    background-color: blue;

[expanded] 
#main  { (user stylesheet)
color: red;
-webkit-border-image: url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAQAAAC1+jfqAAAAiElEQ…Ms+LS30CAhBN5nNxeT5hbJ1zwmji2k+aF6NENIPf/hs54f0sZFUVAMigAAAABJRU5ErkJggg==);

[expanded] 
div  { (user agent stylesheet)
display: block;


iframe style:
[expanded] 
-webkit-background-clip: border-box;
-webkit-background-origin: padding-box;
-webkit-background-size: auto;
background-attachment: scroll;
    body - initial inject-stylesheet-iframe-data.html:4
    /-- overloaded --/ #iframebody - initial user stylesheet
background-clip: border-box;
    body - initial inject-stylesheet-iframe-data.html:4
    /-- overloaded --/ #iframebody - initial user stylesheet
background-color: rgb(0, 128, 0);
    body - green inject-stylesheet-iframe-data.html:4
    /-- overloaded --/ #iframebody - red user stylesheet
background-image: none;
    body - initial inject-stylesheet-iframe-data.html:4
    /-- overloaded --/ #iframebody - initial user stylesheet
background-origin: padding-box;
    body - initial inject-stylesheet-iframe-data.html:4
    /-- overloaded --/ #iframebody - initial user stylesheet
background-size: auto;
    body - initial inject-stylesheet-iframe-data.html:4
    /-- overloaded --/ #iframebody - initial user stylesheet
display: block;
    body - block user agent stylesheet
margin-bottom: 8px;
    body - 8px user agent stylesheet
margin-left: 8px;
    body - 8px user agent stylesheet
margin-right: 8px;
    body - 8px user agent stylesheet
margin-top: 8px;
    body - 8px user agent stylesheet

[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
media="screen" (inject-stylesheet-iframe-data.html)
body  { (inject-stylesheet-iframe-data.html:4)
background: green;
    background-image: initial;
    background-position-x: initial;
    background-position-y: initial;
    background-size: initial;
    background-repeat-x: initial;
    background-repeat-y: initial;
    background-attachment: initial;
    background-origin: initial;
    background-clip: initial;
    background-color: green;

[expanded] 
#iframebody  { (user stylesheet)
/-- overloaded --/ background: red;
    /-- overloaded --/ background-image: initial;
    /-- overloaded --/ background-position-x: initial;
    /-- overloaded --/ background-position-y: initial;
    /-- overloaded --/ background-size: initial;
    /-- overloaded --/ background-repeat-x: initial;
    /-- overloaded --/ background-repeat-y: initial;
    /-- overloaded --/ background-attachment: initial;
    /-- overloaded --/ background-origin: initial;
    /-- overloaded --/ background-clip: initial;
    /-- overloaded --/ background-color: red;

[expanded] 
body  { (user agent stylesheet)
display: block;
margin: 8px;
    margin-top: 8px;
    margin-right: 8px;
    margin-bottom: 8px;
    margin-left: 8px;
